Replace MethodAndVariableSnakeCase with MethodAndVariableNaming.
Add config parameter EnforcedStyle, with values snake_case and camelCase.
