+++ lib/rubocop.rb
@@ -43,6 +43,7 @@ require 'rubocop/cop/lint/useless_setter_call'
 require 'rubocop/cop/lint/void'
 
 require 'rubocop/cop/style/autocorrect_alignment'
+require 'rubocop/cop/style/configurable_naming'
 require 'rubocop/cop/style/string_help'
 require 'rubocop/cop/style/access_control'
 require 'rubocop/cop/style/alias'
@@ -93,10 +94,10 @@ require 'rubocop/cop/style/lambda'
 require 'rubocop/cop/style/lambda_call'
 require 'rubocop/cop/style/leading_comment_space'
 require 'rubocop/cop/style/line_length'
-require 'rubocop/cop/style/method_and_variable_snake_case'
 require 'rubocop/cop/style/method_call_parentheses'
 require 'rubocop/cop/style/method_called_on_do_end_block'
 require 'rubocop/cop/style/method_length'
+require 'rubocop/cop/style/method_name'
 require 'rubocop/cop/style/module_function'
 require 'rubocop/cop/style/multiline_block_chain'
 require 'rubocop/cop/style/multiline_if_then'
@@ -134,6 +135,7 @@ require 'rubocop/cop/style/trailing_whitespace'
 require 'rubocop/cop/style/trivial_accessors'
 require 'rubocop/cop/style/unless_else'
 require 'rubocop/cop/style/variable_interpolation'
+require 'rubocop/cop/style/variable_name'
 require 'rubocop/cop/style/when_then'
 require 'rubocop/cop/style/while_until_do'
 require 'rubocop/cop/style/word_array'
